ok -- change restrict-deformation usage
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@390 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ ImageRegistration/itkANTSImageRegistrationOptimizer.h
@@ -656,7 +656,7 @@ PointSetPointer  WarpMultiTransform(ImagePointer referenceimage, ImagePointer mo
 
 // set up parameters for deformation restriction 
     std::string temp=this->m_Parser->GetOption( "Restrict-Deformation" )->GetValue();
-    this->m_RestrictDeformation = this->m_Parser->template ConvertVector<unsigned int>(temp);
+    this->m_RestrictDeformation = this->m_Parser->template ConvertVector<float>(temp);
     if ( this->m_RestrictDeformation.size() != ImageDimension ) {
       std::cout <<" You input a vector of size :  "  << this->m_RestrictDeformation.size() << " for --Restrict-Deformation.  The vector length does not match the image dimension.  Ignoring.  " << std::endl;
       for (unsigned int jj=0; jj<this->m_RestrictDeformation.size();  jj++ )
@@ -1659,7 +1659,7 @@ private:
   PointSetPointer  m_FixedPointSet;
   PointSetPointer  m_MovingPointSet;
   std::vector<unsigned int> m_Iterations;
-  std::vector<unsigned int> m_RestrictDeformation;
+  std::vector<float> m_RestrictDeformation;
   std::vector<float> m_RoiNumbers;
   float m_GradSmoothingparam;
   float m_TotalSmoothingparam;
